refactor(agents/core): continue making use of operation guard
Carry on previous refactor work moving functionality to the operations modules and
out of the specs.
